Convert to a subcommand approach
The addition of the audit action made clear that the script was going
to get out of hand pretty quickly.  So take the time now to use
picocli's subcommand features and split the various actions into
subcommands.  This will make the addition of new features/actions in
the future much easier.
